Backport (unconditionally) to Evolution 2.32.
It'll make life a lot easier for everyone if we can stop chasing the
bleeding edge and develop on on 2.32 for the time being.

I spent yesterday trying to update to evo git HEAD and make everything work,
which involved local builds of gtk3 and a bunch of other stuff. It fails;
segfaults during startup somewhere in gobject code. This is not an efficient
use of my time, and everyone in the team is blocked until we can all get
it to work.

We're actually going to want to deploy this on 2.32 anyway, and there isn't
much work to do to forward-port to HEAD again either; this patch is
deliberately dirty and leaves the /*cancellable,*/, /*_sync*/ parts there
but commented out, so it can mostly be undone with a simple search/replace.

It *could* also be littered with ifdefs so that it'll build either way, but
for now let's not do that — let's just all build on 2.32. That way we don't
get the 2.32 people breaking the HEAD build and vice versa; we can just
concentrate on making it work.