Fix race conditions with libsoup invocation.
According to https://bugzilla.gnome.org/show_bug.cgi?id=642573 we are
not supposed to invoke soup_session_queue_request() from a thread other
than the main one. So instead of using a signal which gets invoked
directly, add use g_idle_add_full() to trigger the next request to be
queued from the mainloop instead.

Whether that's a sane design for an API is an open question, but the
fact remains that there are existing versions of libsoup which suffer
this problem, and we need to cope.