Basic implementation of mail folder synchronisation.
The Exchange server is fairly screwed and doesn't seem to have a sane
representation of *any* flags other than 'seen', which is implemented by
this patch. It doesn't seem to handle 'forwarded', 'replied', or 'deleted'
in any fashion. Except of course actual deletion, which in the standard
mail setup would be 'expunge'.