Parse some MAPI fields to infer forwarded and answered status for mail.
We have to look at the PidTagIconIndex MAPI property. That field has no
business even *existing*, as far as I can tell. But there seems to be no
better way.

Exchange is so broken in its handling of this stuff that the \Answered
flag from IMAP doesn't even seem to propagate to Outlook, and vice versa.

How people can write shit like this is entirely beyond me.