Stop camel_ews_store_summary_remove_folder() removing wrong name from hash
It may get called with a folder that's *already* had its name removed from
the id->fname lookup hash, during a folder rename. In that case, we must
not remove the *new* name from the hash. Only do the removal if the existing
entry in the hash matches the folder we're removing now.