Fix handling of renamed folders in sync_updated_folders()
This was buggered in various ways, not least of which being the fact that
it was passing the folderid to c_e_s_s_set_folder_name() in place of the
folder's name. But the logic to create the new folder name from the new
parent folder ID (if any) and the new display name (if any) was fairly
broken too. And the folder_name pointer doesn't seem to live long enough
for us to use it to the end of the function; we need a copy of that string
instead.

Add some comments showing what's going on.