Add own thread and mainloop for libsoup.
The Evolution UI will sometimes call our folder rename function synchronously
from the main loop. This is a bug, but for 2.32 at least we need to cope with
it. So start by putting all the libsoup handling into its own thread, so it
doesn't rely on idle callbacks from the main thread.