Remove ParentFolderId from store summary. It's redundant.
The lookup key for the store summary is the full path name. So you have to
already *know* the full path to the folder you're querying, anyway. And
in the case where you're doing a lookup from the FolderId, we have a hash
table for that which gives you the full pathname.

It was also making me sad when I was implementing folder rename, moving
a folder to be a child of the msgfolderroot, and didn't *know* the new
parent folder ID.