Provide simpler ewscal_set_time() function.
There's no point in converting to a time_t just to use gmtime() to convert
it back to yyyy/mm/dd-hh-mm-ss again. Just use the data directly from the
icaltimetype. And have one function which takes the "Start" or "End"
string as a parameter; there's no point in having two functions which do
the same thing.