Don't *require* username to e_ews_connection_authenticate()
Still *allow* it to be given, because it may turn out to be useful in
the autodiscover case. We might want to let them override the username
there, as well as setting the password. But no need to force everyone else
to pass it in when we already know it.