Handle authentication failure in calendar.
No point in using the "authenticate" callback here, since it doesn't look
like we have any option to ask the user for a password on demand anyway.

Instead, just hook it up to trigger an immediate sync on opening the calendar.
Store the context so that we can report success or failure according to the
status of the sync request.

This appears to pass basic testing, although this would be easier if I could
force gnome-keyring to forget the damn password on demand.