Stream MIME content directly to cache when fetching mail messages
Before this, we had *three* copies of the entire message in memory. One in
the SoupBuffer for the response, one in the parsed xmlDoc, and a final
copy in the MimeContent. And on huge messages we were getting out of memory
error from libxml2 on the second of those, resulting in truncated messages.