Revamp store summary to be keyed on folder_id not full_name
We shouldn't be keeping the full name in disk storage at all, let alone
using it as the key for the database. It's secondary information which can
be recreated from the *real* data, which is the individual folder names
and the parent folder ID for each folder.

Build up hash tables for id->fullname and fullname->id lookups, and change
all callers to deal with the store-summary by folder ID instead of by name.