Remove duplicate folder_info_build() function
We already had camel_ews_utils_build_folder_info(). Now, *perhaps* that
function wasn't doing the right thing somehow. But if that were the case
then the new function would have had an appropriate name, and
appropriate comments both in the code, and in the commit comment. In the
absence of any such markers, we have to assume that it's purely an
oversight that the original function wasn't used. (I'd test it, but this
code is still too broken to do even basic testing on it yet).

Perhaps it wasn't obvious that someone, somewhere, would already have
written some code to create a CamelFolderInfo for an EWS folder?

Gr.