Fix broken error handling in create/delete folder.
The 'if (error) return NULL;' construct is doubly broken.

The 'GError **error' parameter will almost *always* be non-NULL, since
it's pointing to a local 'GError *' variable somewhere in our caller (or
further up in the call stack). So we were returning NULL here *any* time
that we were called by a caller that wanted to get detailed error
information.

Only if our caller passed NULL as the error pointer, because it didn't
care about error details, did we ever get past this point! It makes no
difference whether the call to e_ews_connection_*_folder() actually
succeeded or not. So it looks like this code was not even *slightly*
tested. Gr.

There are two variations on this broken theme which have been seen in
the wild, and which I thought I had already eliminated from the evo-ews
code base:

Firstly, there's 'if (*error)'. That will SEGV if the caller didn't care
about errors and just passed us NULL.

Then there's 'if (error && *error)'. That won't crash, at least, but it
*will* miss errors, if an error happened but the caller passed NULL because
it didn't want the details.

We need to actually check the return value from the function, not look at
the error.

Gr.