Fix various issues with folder creation in ews_get_folder_sync()
 - 'flags' is a bitfield, and should be checked as such rather than with ==
 - memory leak of 'parent' string if folder did already exist
 - memory leak on the new call to c_e_s_s_get_folder_id_from_name() if exists
 - overwriting already-set error if ews_create_folder_sync() fails
 - using strdup/free instead of g_strdup/g_free
 - memory leak of CamelFolderInfo returned by ews_create_folder_sync()
 - cosmetic coding style issues (no space before function args, mostly)

Also fixed a pre-existing potential overwrite of an error returned by
camel_ews_folder_new(), and free 'fid' earlier to make it clear we don't
actually *need* it; we only wanted to know that it exists. And only strdup
the folder_name if we actually need to, which isn't strictly a bug fix; it
just makes me slightly happer. Kernel hacker mentality :)