Simple fix for crash on GetItems
On some items, message:From is invalid and Exchange returns an error
(instead of just returning the details which *are* valid, along with a
warning. FPOS).

Before, this would have caused the whole GetItems request to fail, and it
looks like we were ignoring that error. Ignoring the error is a bug in itself,
which needs to be fixed.

And also, we need to stop requesting properties that don't exist, given
how badly Exchange handles that. So either we need to request the
Default shape instead of using IdOnly and adding a list of props, or we
need to keep track of which items are of which type and request the
*appropriate* properties for each.