Revert all the EEwsOperation changes.
We'll re-use *some* of this code, but for now let's start again from scratch.

In particular, we don't want separate subclasses for *every* operation. We
want an EEwsOperation class, but *all* operations should have that class.

We want a 'create request' function for each server operation, which
returns an EEwsOperation with the appropriate SoapMessage content.

The caller will:
 - Call the appropriate 'create request' function.
 - Complete the request SoapMessage if necessary (replacing the create_cb
   callbacks for UpdateItem etc.)
 - Set any other optional parameters like progress_fn, streaming to file
   (replacing the progress_fn and similar arguments which were causing the
    e_ews_connection_get_items_start() and similar function prototypes to
    grow without bounds).
 - Override the default response parsing callback, if desired.
 - Submit the message, asynchronously or synchronously.