Revert broken Windows build fix
This calls a function that only exists in e-util.so. But you can't do that
from the back ends; that's an Evolution GUI library and doesn't exist in
e-calendar-factory and e-addressbook-factory.

Conversely, we *can* link to camel libraries from the account setup
plugin (at least for now). And we do. So removing the linkage results
in:
(evolution:29894): evolution-plugin-lib-WARNING **: can't load plugin '/usr/lib64/evolution/3.0/plugins/liborg-gnome-exchange-ews.so': /usr/lib64/evolution/3.0/plugins/liborg-gnome-exchange-ews.so: undefined symbol: camel_ews_store_summary_remove