evolution-ews: implement Exchange categories as evolution label
This patch is a proof of concept for adding evolution labels to an
exchange back end.  The outlook categories perform roughly the same
function, and evolution even mirrors their colour scheme.

This patch allows evolution to fill in the labels from the EWS
Category element.  That means that any colour you add to a message via
the categories in outlook (or the outlook web server) automatically
appears in evolution.

The category modification is done from the synchronize backends and
simply replaces all of the item categories with the camel ones.

Signed-off-by: James Bottomley <James.Bottomley@parallels.com>