evolution-ews: fix draft flag on sent items
What's happening currently is that the outlook web server shows all my
sent messages with a yellow banner wrongly saying "This Message has
not been Sent".

The design of EWS is that you should append to folders with
MessageDisposition "SaveOnly" and send with "SendAndSaveCopy".  The
former always sets the IsDraft flag and the latter always clears it.
Unfortunately, the design of evolution precludes us from using
SendAndSaveCopy since evolution wants to send and then separately save
the sent email.  Thus we need to modify SaveOnly to respect the
CAMEL_MESSAGE_DRAFT flag by setting the MAPI message flags.

I also removed the FIXME: comment about the From: field.  This is
irrelevant to us, since EWS respects all the properties in the MIME
body (which include the From: field and the Importance: field).

Signed-off-by: James Bottomley <JBottomley@Parallels.com>