evolution-ews: allow set/reset of Importance flag
When you have a lot of people who set the Importance flag to high on
all their incoming messages, it's useful to be able to reset it so
that only genuinely important messages remain flagged in your folders.

This patch implements the ability to set the Importance flag from
evolution.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>