evolution-ews: fix calendar failure when specifying UTC as the TimeZone
If you use evolution to specify an appointment in UTC, currently it
gets translated as native time in whatever timezone the exchange
server happens to be running.

The bug is that the UTC timezone has no properties, so the check for
standard properties in ewscal_set_timezone() fails.  Fix by assuming
that if the timezone has no properties, it must be UTC.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>