Kill X-EVOLUTION-ACCEPT-ID field; we just use X-EVOLUTION-ITEM-ID
It's not clear how the receive_objects code will ever cope if we are asked
to accept an item that *didn't* have an AssociatedCalendarItemId; the old
code was just using the UID in that case, but I think that was expected to
fail. Do the same for now... weren't we working on fixing this?