evolution-ews: fix timezone issue on copying from calendars
Currently, when you copy between calendars, it's hit or miss whether
the appointment copies to the correct time.  The issue is whether the
icaltimezone gets resolved or not.  Invariably, if the tzid is
non-standard, it doesn't.

The fix is to always resolve the tzid to a vtimezone component in the
current calendar and append that to the icalcomponent.  The fix
depends on the fact that the resolution of the tzid is in the current
calendar, which depends on fixing

https://bugzilla.gnome.org/show_bug.cgi?id=655843

Signed-off-by: James Bottomley <JBottomley@Parallels.com>