Don't ref/unref connection repeatedly in calendar code.
We hold a ref on it via cbews->priv, so this shouldn't be needed.

I suspect this was causing the stray unref of cnc which caused problems
when we fixed the GSimpleAsyncResult leak. Before we fixed the leak, those
leaked objects would each have a reference on the connection and would
mask the problem.