Reinstate leak fix for GSimpleAsyncResult objects
Now that the stray connection unref should be fixed, we can reinstate
the fix for the leaked GSimpleAsyncResults. A slightly improved version
over commit 1ac06aa5, this one prevents a possible use-after-free
because the connection is pinned by the result, so we must do it *in*
ews_active_job_done() after we've done everything we want to with
enode->cnc.