Rework SOAP response handling.
ews_response_cb() overgeneralizes and tries to do too much for too many
different kinds of SOAP responses.  This results in confusing response
handling and error handling.

Simplify things by just handing the whole ESoapResponse to the response
callback function and let it process it.  This approach results in some
repetitive logic for handling ResponseMessage elements in 20 different
places, but the logic is simple and straightforward.