Fix crash on addressbook close
In e_book_backend_ews_dispose() we cancel priv->cancellable and then
immediately free it along with a bunch of other stuff, and *then* wait
for the update thread to complete.

The update thread then crashes the next time it checks if the (now-freed)
cancellable is cancelled.

We should first cancel it, *then* wait for the thread, and then free
everything else. In fact I'm not entirely sure we *should* be freeing
everything else here; isn't that supposed to be done in the finalize()
method? But I'll leave it as-is for now...
(cherry picked from commit a56a32e1044a7932e91425a0398a0ae00c9398a5)