Revert "Bug #703181 - Book/calendar backends can authenticate without password"
This reverts commit addcca8b5803a3abcb532f52523962fae3fa0e11.

The added ews_connection_schedule_abort (connection); at
ews_connection_try_password_sync() is not enough, the second attempt
will not work and fails to connect on the session, thus let's rather
revert this for now, because it's too early before the release, and
revisit this later. The associated API change in evolution-data-server
will not be reverted.