Bug 703181 - Work around libsoup's failure to recover from NTLM auth failure
Left to its own devices, libsoup will end up with NTLM authentication
permanently in SOUP_NTLM_FAILED state, even if we give it a new password.

Even without winbind/ntlm_auth in the picture, this was a problem. If a
user mistyped their password once, it wouldn't be accepted on the second
attempt even if they did get it right.

With automatic NTLM authentication being attempted, this failed state was
happening even more frequently. Hence the reversion of commit addcca8 when
it was noticed (see bug 703181 comments 20-21).

Proper fixes are going into libsoup, of course, but this serves to work
around it for now by cancelling the message when we can't authenticate.