Bug 732850 - Remove reachability check from ews_store_query_auth_types_sync()
The auth method discovery code first checks whether the "host" is
reachable. This will always fail because it only sets its Host URL for
the service, and the camel code relies on the host and port itself being
set before it can answer that question.  Even after setting the host and
port, it is necessary to give the underlying system time to query the
network service.  So this check should not be done at all.

[dwmw2: The check is entirely redundant anyway. Even after we rip it out,
        I've confirmed that we still get coherent errors reported in the
        UI, telling us that we were unabled to resolve or reach the server.]