Add camel_ews_settings_get_auth_mechanism() helper function
The auth-mechanism string is not fun to deal with. It can sometimes
be NULL and sometimes an empty string for NTLM, and I've managed to
screw up even the *simple* cases, comparing for "BASIC" and then "Basic"
in subsequent broken commits (b6ff74a4, 377da623) instead of "PLAIN".

Make it nice and simple. Hard-code this knowledge about the strings in
*one* place, camel_ews_settings_get_auth_mechanism(), and return an
enum. Which in turn will help us to ensure we handle all cases, in
places where that's appropriate (such as in
e_ews_connection_utils_get_without_password())