SOUP_STATUS_IS_SUCCESSFUL(soup_session_send_message()) sends message twice
The definition of SOUP_STATUS_IS_SUCCESSFUL(status) is
	((status) >= 200 && (status) < 300)

So if we actually pass the function call as the 'status' parameter of
the macro, it's going to get called twice, and thus the message will
be submitted twice.

This is obviously suboptimal, and it actually breaks with the GSSAPI
 hacks, because the second submission just keeps the same Authorization:
header that was on the message last time, and that's seen as a replay
attack by the server.

(cherry picked from commit aa4f35dee7b04f7f858c17e728cbd6b0fd938984)