Pass GInputStream to decoder functions instead of assuming priv->fis
If we want to optimise the decode a little and also use data twice, we'll
probably want to create a GMemoryInputStream for each record. We're doing
a lot of seeking due to ews_oab_read_upto() already, and it'll only get
worse if we take the SHA1 of each record as well as processing it.

(cherry picked from commit 140d2ba3ad2d27193119de4b81d0b92c09204868)