Pass GInputStream to decoder functions instead of assuming priv->fis
If we want to optimise the decode a little and also use data twice, we'll
probably want to create a GMemoryInputStream for each record. We're doing
a lot of seeking due to ews_oab_read_upto() already, and it'll only get
worse if we take the SHA1 of each record as well as processing it.