Optimise EWS GAL update not to rewrite unchanged records
Store the SHA1 of the binary data in bdata in the sqlite. When processing
the new OAB file, if the bdata already matches then do nothing to the
existing record.

Before starting to process the OAB file, build a GHashTable with all the
UIDs which already exist in the database. Remove them from the GHashTable
as they are seen while processing the OAB file. At the end, any which are
still left in the GHashTable are deleted.

The value in the GHashTable is the SHA1, and is freed either when a
matching record is decoded or at the end when we process the hash table
to remove all remaining entries, with the append_to_list() helper
function.

This should *drastically* improve the performance of the GAL update.

(cherry picked from commit f0e4e8cd6bbf957b1952efe1c7519512388a0b84)
[Change message string back to avoid string changes]