Avoid processing OAB records which already exist
By checking the SHA1 before generating the EContact we can halve the time
taken to process the Intel GAL from about 50s to about 25s.

The extra data is now the key in one GHashTable, and freed with that
hash table. While the uids are the key in the other and freed with that.

(cherry picked from commit 1a493d97cc447a7406e70ab5e83427233e1fc48a)