Make ebews_fetch_items() return a list of EContacts not EVCards
... and fix up e_book_backend_ews_get_contact_list() to cope.

We'll actually want to ditch the store_to_cache mode of ebews_fetch_items()
and *always* use it to return EContacts, even from ews_update_items_thread().
That way we can do all the network I/O in advance and then do the database
update in a single transaction.

Fix a crash in the database-query code paths in get_contact_list() too while
we're at it. On converting to EBookSqlite, the 'list' variable needed to be
initialised to NULL.