Synchronize authentication extensions before creating an ESource
When the values in the collection source and in the child sources
of the authentication extension do not match, then the GUI part can
ask for the password even when not needed. Even the collection-backend
listens for child-added, it's too late to fix it there, because other
processes can be notified about these new sources as well and use
old values.