I#81 - Attachment flag not shown on messages with attachment ][
Extends the previous change to:
a) set/unset the attachment flag after the message is loaded from the server;
b) add a guess on common content types to not set the attachment flag in error.

Even if the b) makes it wrong, the a) can correct it later.

Related to https://gitlab.gnome.org/GNOME/evolution-ews/issues/81