Mark connection disconnected only when disconnecting backend/store
The connection had been marked as disconnected, thus unusable for any
new attempts to connect to the same server, both when backend/store
had been disconnected and when it had been freed. The later case
is wrong, because there was no problem with the connection, thus
it can be used by other backends (when multiple calendars/books are
configured).