Add (disabled by default) Microsoft 365 (Graph API) connector
To enable the connector users need to export ENABLE_M365=1 environment
variable in a way that it's known both to the Evolution itself and
to the evolution-data-server background services. One such place can
be in /etc/environment.

It's disabled by default, because the Microsoft Graph API is currently
far from being feature complete with the EWS API, thus it's here only
as kind of a preview.