Listen for change notifications only on user folders
The foreign and public folders cannot be used for change notifications,
due to the server returning and error "A subscription can only be established
on a single public folder or on folders from a single mailbox.", thus, rather
than not having any change notifications, do listen on user folders only.
The mail part did that already, only address book and calendar parts
did not do it.