Prefill compile-time OAuth2 settings, to be shown in the GUI
The compile-time settings had been used in the background, but they
were not visible in the GUI, only as part of the tooltip on respective
widgets and it might not be clear that the values can be left (visually)
empty. The only downside is that the values will be stored in the memory
and in the files, but it might not be a big deal.