EEwsConnection: try_credentials() may end with 'Forbidden' error
When multiple sources share the same EWS connection object, they
can ask to try credentials at the same time, which then fights
on the password usage and cached credentials. This makes sure
to not do that.

Also report the 'Forbidden' response with no password provided
as an authentication error, thus ask for credentials.