Microsoft365: Implement raw MIME message upload
The message is still marked as draft (isDraft flag) on the server.
There does not seem to be a way to unset it. There also does not work
upload to the specified folder, the server errors out, thus the code
uploads to the Drafts folder and then moves the message to the specified
folder. Maybe that plays its role with the isDraft flag as well.