EWS calendar: Do not link against Evolution libraries
Replace Evolution functions with newly added libecal functions
and copy/adapt the rest Evolution functions, to not link against
Evolution, thus to not load Evolution (and GUI) libraries into
the evolution-calendar-factory process.

The code duplication is not ideal, but it does worth it in this case.

This way the distribution packagers can split evolution-ews into packages
which require evolution (the configuration part) and which not, and let
the users install evolution-ews core parts without Evolution itself, thus
for example Exchange accounts configured in GNOME Online Accounts can
provide data to GNOME Calendar without a need to install Evolution itself.