m365 Camel: Auto-save store summary on folder count changes
The folder counts can come in any time, without the code knowing when the change
will be saved, thus schedule a timeout to save the changes, thus they are not lost.
This could be mainly visible when the unread count changes on the folder without
being saved, thus the next start the Evolution indicated new messages had been
received to the folder, even though there was no new messages at all.