m365 Camel: Send messages as MIME data
The internal conversion from MIME to Json can miss some important information,
like currently for example thread references, thus rather pass the message
as a MIME blob than trying to fill related properties of a Json object.

A disadvantage is that the messages are always stored to the Sent Items folder.