Misc: Extract common code into a new shared library
The are functions copied from the Evolution code, to not depend on Evolution
in the backends and modules, which could bring into the random processes also
the gtk dependency, which could break when the app was not using the same
gtk version as Evolution.

This limits the code duplication at least within the evolution-ews.